+++ packages/es-dev-server/demo/babel/server.js
@@ -1,0 +1,8 @@
+module.exports = {
+  rootDir: '../../',
+  appIndex: 'packages/es-dev-server/demo/babel/index.html',
+  nodeResolve: true,
+  babel: true,
+  open: true,
+  watch: true,
+};