+++ packages/es-dev-server/demo/performance/server.js
@@ -1,0 +1,7 @@
+module.exports = {
+  rootDir: '../../',
+  appIndex: 'packages/es-dev-server/demo/performance/index.html',
+  nodeResolve: true,
+  open: true,
+  watch: true,
+};