+++ packages/rollup-plugin-index-html/test/snapshots/legacy/lazy-a8a6feca.js
@@ -1,3 +1,0 @@
-import './shared-6e98f23f.js';
-
-console.log('my lazy');