+++ packages/rollup-plugin-index-html/test/snapshots/single-entry/lazy-1d008aa1.js
@@ -1,0 +1,3 @@
+import './shared-ed942ddb.js';
+
+console.log('my lazy');